requirements.txt
Pillow==11.2.1
Requests==2.32.3
UnityPy==1.21.0
astc-encoder-py==0.1.11


